# Improve LINQ predicate performance