--- a/src/System.Linq/src/System/Linq/Single.cs
+++ b/src/System.Linq/src/System/Linq/Single.cs
@@ -138,10 +138,9 @@ namespace System.Linq
             }
             else if (source is List<TSource> list)
             {
-                TSource element;
                 for (int i = 0; i < list.Count; ++i)
                 {
-                    element = list[i];
+                    TSource element = list[i];
 
                     if (predicate(element))
                     {