Fixes #13696 and #39148 

This PR adds type checks and code paths for List<T> and T[] to `First(predicate)`, `FirstOrDefault(predicate)`, `Single(predicate)` and `SingleOrDefault(predicate)`.

`Any(predicate)` now uses the `TryGetFirst` method, I thought this might be better than doing the same type checks in there as well.

Things that might need some fixing still:
- ~~TryGetSingle has a `out bool found` and `out bool foundSingle` because I didn't think that throwing an exception from a method prefixed with 'Try' would be acceptable.~~ Fixed.
- ~~Might need to separate things with a `SpeedOpt` or `SizeOpt` implementation as well~~ This appears to be about memory usage, and not code size (?).

Benchmarks are submitted to dotnet/performance [here](https://github.com/dotnet/performance/pull/611).