@jvandertil 2019-07-04
Before (against master):
``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=Windows 10.0.17763.593 (1809/October2018Update/Redstone5)
Intel Core i7-7700K CPU 4.20GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview6-012264
  [Host]     : .NET Core 3.0.0-preview6-27804-01 (CoreCLR 4.700.19.30373, CoreFX 4.700.19.30308), 64bit RyuJIT
  Job-OGFUHE : .NET Core ? (CoreCLR 4.700.19.35202, CoreFX 4.700.19.35401), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                                          Method | size | iteration |            wrapType | iterationCount |            Mean |          Error |         StdDev |          Median |             Min |             Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------------------ |----- |---------- |-------------------- |--------------- |----------------:|---------------:|---------------:|----------------:|----------------:|----------------:|------------:|------------:|------------:|--------------------:|
|                   **WhereFirst_LastElementMatches** |  **100** |         **?** |              **NoWrap** |           **1000** |       **232.38 ns** |      **1.5346 ns** |      **1.2815 ns** |       **232.31 ns** |       **230.57 ns** |       **234.99 ns** |      **0.0319** |           **-** |           **-** |               **136 B** |
|           FirstWithPredicate_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       499.60 ns |      9.4758 ns |      9.3065 ns |       496.25 ns |       490.93 ns |       517.56 ns |      0.0281 |           - |           - |               120 B |
|          WhereFirstOrDefault_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       227.68 ns |      0.7051 ns |      0.6251 ns |       227.52 ns |       226.80 ns |       229.01 ns |      0.0321 |           - |           - |               136 B |
|  FirstOrDefaultWithPredicate_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       518.47 ns |      4.1090 ns |      3.8435 ns |       516.66 ns |       514.43 ns |       525.88 ns |      0.0270 |           - |           - |               120 B |
|                     WhereAny_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       216.41 ns |      2.8273 ns |      2.6446 ns |       215.48 ns |       214.00 ns |       222.73 ns |      0.0321 |           - |           - |               136 B |
|             AnyWithPredicate_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       522.35 ns |      8.7572 ns |      8.1914 ns |       523.39 ns |       513.87 ns |       536.70 ns |      0.0270 |           - |           - |               120 B |
|                  WhereSingle_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       250.84 ns |      2.8812 ns |      2.6950 ns |       249.56 ns |       248.17 ns |       256.59 ns |      0.0318 |           - |           - |               136 B |
|          SingleWithPredicate_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       497.39 ns |      3.1423 ns |      2.6240 ns |       498.16 ns |       492.56 ns |       500.30 ns |      0.0278 |           - |           - |               120 B |
|         WhereSingleOrDefault_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       251.21 ns |      0.5130 ns |      0.4284 ns |       251.23 ns |       250.57 ns |       252.03 ns |      0.0322 |           - |           - |               136 B |
| SingleOrDefaultWithPredicate_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       519.95 ns |      5.7122 ns |      5.3432 ns |       516.83 ns |       515.21 ns |       530.04 ns |      0.0268 |           - |           - |               120 B |
|                                       WhereLast |  100 |         ? |              NoWrap |           1000 |       386.93 ns |      3.7178 ns |      3.4777 ns |       385.17 ns |       384.23 ns |       392.94 ns |      0.0108 |           - |           - |                48 B |
|                               LastWithPredicate |  100 |         ? |              NoWrap |           1000 |        29.63 ns |      0.2745 ns |      0.2568 ns |        29.70 ns |        29.29 ns |        30.21 ns |           - |           - |           - |                   - |
|                              WhereLastOrDefault |  100 |         ? |              NoWrap |           1000 |       398.24 ns |      4.3688 ns |      4.0865 ns |       396.55 ns |       394.77 ns |       408.16 ns |      0.0112 |           - |           - |                48 B |
|                      LastOrDefaultWithPredicate |  100 |         ? |              NoWrap |           1000 |        28.69 ns |      0.2837 ns |      0.2653 ns |        28.58 ns |        28.40 ns |        29.08 ns |           - |           - |           - |                   - |
|                        Contains_ElementNotFound |  100 |         ? |              NoWrap |           1000 |        44.21 ns |      0.1282 ns |      0.1001 ns |        44.19 ns |        44.11 ns |        44.42 ns |           - |           - |           - |                   - |
|                    Contains_FirstElementMatches |  100 |         ? |              NoWrap |           1000 |        23.19 ns |      0.1882 ns |      0.1761 ns |        23.26 ns |        22.91 ns |        23.49 ns |           - |           - |           - |                   - |
|                   **WhereFirst_LastElementMatches** |  **100** |         **?** |         **IEnumerable** |           **1000** |       **613.99 ns** |      **6.5454 ns** |      **6.1225 ns** |       **612.54 ns** |       **606.85 ns** |       **622.41 ns** |      **0.0465** |           **-** |           **-** |               **200 B** |
|           FirstWithPredicate_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       541.87 ns |      1.1398 ns |      0.8899 ns |       541.70 ns |       540.56 ns |       543.13 ns |      0.0325 |           - |           - |               144 B |
|          WhereFirstOrDefault_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       576.25 ns |      3.5964 ns |      3.3641 ns |       574.10 ns |       572.98 ns |       581.61 ns |      0.0460 |           - |           - |               200 B |
|  FirstOrDefaultWithPredicate_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       543.73 ns |      1.0184 ns |      0.7951 ns |       543.45 ns |       542.88 ns |       544.91 ns |      0.0326 |           - |           - |               144 B |
|                     WhereAny_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       582.60 ns |      5.8311 ns |      5.4544 ns |       579.17 ns |       577.48 ns |       591.13 ns |      0.0464 |           - |           - |               200 B |
|             AnyWithPredicate_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       542.22 ns |      7.5639 ns |      6.7052 ns |       540.08 ns |       534.14 ns |       552.48 ns |      0.0323 |           - |           - |               144 B |
|                  WhereSingle_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       592.87 ns |      6.2300 ns |      5.8276 ns |       589.42 ns |       587.35 ns |       603.21 ns |      0.0469 |           - |           - |               200 B |
|          SingleWithPredicate_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       562.75 ns |      5.9027 ns |      5.5214 ns |       565.19 ns |       555.07 ns |       570.13 ns |      0.0334 |           - |           - |               144 B |
|         WhereSingleOrDefault_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       594.84 ns |      0.9326 ns |      0.7281 ns |       594.82 ns |       593.74 ns |       596.03 ns |      0.0477 |           - |           - |               200 B |
| SingleOrDefaultWithPredicate_LastElementMatches |  100 |         ? |         IEnumerable |           1000 |       560.09 ns |      3.6951 ns |      3.4564 ns |       561.53 ns |       555.03 ns |       564.62 ns |      0.0335 |           - |           - |               144 B |
|                                       WhereLast |  100 |         ? |         IEnumerable |           1000 |       765.74 ns |      1.8027 ns |      1.4074 ns |       765.51 ns |       763.70 ns |       768.89 ns |      0.0245 |           - |           - |               112 B |
|                               LastWithPredicate |  100 |         ? |         IEnumerable |           1000 |       519.23 ns |      5.9633 ns |      5.2863 ns |       516.33 ns |       512.99 ns |       526.37 ns |      0.0124 |           - |           - |                56 B |
|                              WhereLastOrDefault |  100 |         ? |         IEnumerable |           1000 |       748.70 ns |      4.4926 ns |      3.9826 ns |       747.27 ns |       745.29 ns |       759.97 ns |      0.0240 |           - |           - |               112 B |
|                      LastOrDefaultWithPredicate |  100 |         ? |         IEnumerable |           1000 |       551.21 ns |      3.3082 ns |      3.0945 ns |       551.66 ns |       547.62 ns |       558.58 ns |      0.0132 |           - |           - |                56 B |
|                   **WhereFirst_LastElementMatches** |  **100** |         **?** | **IReadOnlyCollection** |           **1000** |       **650.28 ns** |     **35.0125 ns** |     **40.3205 ns** |       **636.80 ns** |       **609.99 ns** |       **742.37 ns** |      **0.0458** |           **-** |           **-** |               **200 B** |
|           FirstWithPredicate_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       542.52 ns |      1.5172 ns |      1.1846 ns |       542.37 ns |       540.74 ns |       545.12 ns |      0.0326 |           - |           - |               144 B |
|          WhereFirstOrDefault_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       576.85 ns |      3.3220 ns |      3.1074 ns |       575.87 ns |       572.73 ns |       581.23 ns |      0.0470 |           - |           - |               200 B |
|  FirstOrDefaultWithPredicate_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       540.62 ns |      0.6928 ns |      0.5785 ns |       540.65 ns |       539.64 ns |       541.54 ns |      0.0325 |           - |           - |               144 B |
|                     WhereAny_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       583.89 ns |      6.5037 ns |      6.0836 ns |       579.87 ns |       577.45 ns |       594.71 ns |      0.0464 |           - |           - |               200 B |
|             AnyWithPredicate_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       537.36 ns |      3.6537 ns |      3.4177 ns |       535.88 ns |       534.42 ns |       545.52 ns |      0.0344 |           - |           - |               144 B |
|                  WhereSingle_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       594.15 ns |      3.3606 ns |      3.1435 ns |       594.83 ns |       588.11 ns |       598.67 ns |      0.0473 |           - |           - |               200 B |
|          SingleWithPredicate_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       555.27 ns |      1.3329 ns |      1.1816 ns |       555.13 ns |       553.71 ns |       557.14 ns |      0.0334 |           - |           - |               144 B |
|         WhereSingleOrDefault_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       602.90 ns |      7.7024 ns |      7.2048 ns |       606.32 ns |       594.65 ns |       616.42 ns |      0.0477 |           - |           - |               200 B |
| SingleOrDefaultWithPredicate_LastElementMatches |  100 |         ? | IReadOnlyCollection |           1000 |       556.36 ns |      1.0915 ns |      0.9114 ns |       556.19 ns |       554.50 ns |       557.79 ns |      0.0334 |           - |           - |               144 B |
|                                       WhereLast |  100 |         ? | IReadOnlyCollection |           1000 |       767.95 ns |      3.5180 ns |      3.2907 ns |       766.95 ns |       763.89 ns |       774.25 ns |      0.0245 |           - |           - |               112 B |
|                               LastWithPredicate |  100 |         ? | IReadOnlyCollection |           1000 |       516.00 ns |      2.4682 ns |      2.0611 ns |       515.00 ns |       513.44 ns |       520.83 ns |      0.0124 |           - |           - |                56 B |
|                              WhereLastOrDefault |  100 |         ? | IReadOnlyCollection |           1000 |       752.15 ns |      7.3696 ns |      6.8935 ns |       748.47 ns |       744.40 ns |       762.70 ns |      0.0239 |           - |           - |               112 B |
|                      LastOrDefaultWithPredicate |  100 |         ? | IReadOnlyCollection |           1000 |       549.97 ns |      3.7775 ns |      3.3486 ns |       548.77 ns |       546.77 ns |       558.38 ns |      0.0132 |           - |           - |                56 B |
|                   **WhereFirst_LastElementMatches** |  **100** |         **?** |         **ICollection** |           **1000** |       **608.66 ns** |      **1.3074 ns** |      **1.0917 ns** |       **608.68 ns** |       **606.74 ns** |       **610.69 ns** |      **0.0463** |           **-** |           **-** |               **200 B** |
|           FirstWithPredicate_LastElementMatches |  100 |         ? |         ICollection |           1000 |       548.26 ns |      6.4671 ns |      6.0493 ns |       551.47 ns |       540.07 ns |       557.21 ns |      0.0326 |           - |           - |               144 B |
|          WhereFirstOrDefault_LastElementMatches |  100 |         ? |         ICollection |           1000 |       577.90 ns |      6.3439 ns |      5.9341 ns |       574.76 ns |       573.13 ns |       591.32 ns |      0.0461 |           - |           - |               200 B |
|  FirstOrDefaultWithPredicate_LastElementMatches |  100 |         ? |         ICollection |           1000 |       542.50 ns |      3.0989 ns |      2.8987 ns |       541.26 ns |       538.11 ns |       546.80 ns |      0.0327 |           - |           - |               144 B |
|                     WhereAny_LastElementMatches |  100 |         ? |         ICollection |           1000 |       580.22 ns |      2.0584 ns |      1.6071 ns |       579.96 ns |       577.50 ns |       583.68 ns |      0.0460 |           - |           - |               200 B |
|             AnyWithPredicate_LastElementMatches |  100 |         ? |         ICollection |           1000 |       538.17 ns |      4.7987 ns |      4.4887 ns |       535.70 ns |       534.06 ns |       546.83 ns |      0.0344 |           - |           - |               144 B |
|                  WhereSingle_LastElementMatches |  100 |         ? |         ICollection |           1000 |       593.05 ns |      6.1689 ns |      5.7704 ns |       589.82 ns |       588.28 ns |       603.77 ns |      0.0473 |           - |           - |               200 B |
|          SingleWithPredicate_LastElementMatches |  100 |         ? |         ICollection |           1000 |       558.50 ns |      3.4429 ns |      3.0521 ns |       556.54 ns |       555.53 ns |       563.54 ns |      0.0340 |           - |           - |               144 B |
|         WhereSingleOrDefault_LastElementMatches |  100 |         ? |         ICollection |           1000 |       595.84 ns |      1.2015 ns |      1.1239 ns |       595.62 ns |       593.24 ns |       597.52 ns |      0.0477 |           - |           - |               200 B |
| SingleOrDefaultWithPredicate_LastElementMatches |  100 |         ? |         ICollection |           1000 |       559.01 ns |      4.5449 ns |      4.2513 ns |       556.39 ns |       554.50 ns |       566.80 ns |      0.0334 |           - |           - |               144 B |
|                                       WhereLast |  100 |         ? |         ICollection |           1000 |       769.40 ns |      6.1646 ns |      5.7663 ns |       766.91 ns |       764.18 ns |       781.42 ns |      0.0246 |           - |           - |               112 B |
|                               LastWithPredicate |  100 |         ? |         ICollection |           1000 |       515.02 ns |      1.1291 ns |      0.9429 ns |       514.74 ns |       513.98 ns |       516.82 ns |      0.0124 |           - |           - |                56 B |
|                              WhereLastOrDefault |  100 |         ? |         ICollection |           1000 |       748.58 ns |      2.9213 ns |      2.4394 ns |       748.00 ns |       744.84 ns |       754.60 ns |      0.0239 |           - |           - |               112 B |
|                      LastOrDefaultWithPredicate |  100 |         ? |         ICollection |           1000 |       548.81 ns |      1.4461 ns |      1.2075 ns |       548.72 ns |       546.73 ns |       550.96 ns |      0.0132 |           - |           - |                56 B |

After:
``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=Windows 10.0.17763.593 (1809/October2018Update/Redstone5)
Intel Core i7-7700K CPU 4.20GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview6-012264
  [Host]     : .NET Core 3.0.0-preview6-27804-01 (CoreCLR 4.700.19.30373, CoreFX 4.700.19.30308), 64bit RyuJIT
  Job-DPZQVU : .NET Core ? (CoreCLR 4.700.19.35202, CoreFX 4.700.19.35401), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                                          Method | size | iteration |            wrapType | iterationCount |            Mean |         Error |        StdDev |          Median |             Min |             Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------------------ |----- |---------- |-------------------- |--------------- |----------------:|--------------:|--------------:|----------------:|----------------:|----------------:|------------:|------------:|------------:|--------------------:|
|                   **WhereFirst_LastElementMatches** |  **100** |         **?** |              **NoWrap** |           **1000** |       **208.06 ns** |     **1.4728 ns** |     **1.3777 ns** |       **207.57 ns** |       **206.70 ns** |       **210.78 ns** |      **0.0319** |           **-** |           **-** |               **136 B** |
|           FirstWithPredicate_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       175.61 ns |     1.1656 ns |     1.0903 ns |       175.26 ns |       174.59 ns |       177.52 ns |      0.0206 |           - |           - |                88 B |
|          WhereFirstOrDefault_LastElementMatches |  100 |         ? |              NoWrap |           1000 |       208.60 ns |     1.8414 ns |     1.7224 ns |       207.79 ns |       206.53 ns |       212.00 ns |      0.0317 |           - |           - |               136 B |
…